backend/requirements.txt
astroid==1.6.1
Django==2.0.2
djangorestframework==3.7.7
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.3
six==1.11.0
wrapt==1.10.11
